+++ MapboxDirectionsTests/RouteOptionsTests.swift
@@ -3,8 +3,7 @@ import XCTest
 
 class RouteOptionsTests: XCTestCase {
     func testCoding() {
- 
-        let options = RouteOptions.testInstance
+        let options = testRouteOptions
         let encodedData = NSMutableData()
         let keyedArchiver = NSKeyedArchiver(forWritingWith: encodedData)
         keyedArchiver.requiresSecureCoding = true
@@ -18,7 +17,7 @@ class RouteOptionsTests: XCTestCase {
         
         XCTAssertNotNil(unarchivedOptions)
         
-        let coordinates = RouteOptions.testCoordinates
+        let coordinates = testCoordinates
         let unarchivedWaypoints = unarchivedOptions.waypoints
         XCTAssertEqual(unarchivedWaypoints.count, coordinates.count)
         XCTAssertEqual(unarchivedWaypoints[0].coordinate.latitude, coordinates[0].latitude)
@@ -35,8 +34,40 @@ class RouteOptionsTests: XCTestCase {
         XCTAssertEqual(unarchivedOptions.includesVisualInstructions, options.includesVisualInstructions)
         XCTAssertEqual(unarchivedOptions.roadClassesToAvoid, options.roadClassesToAvoid)
     }
+    
+    func testCodingV4() {
+        let options = testRouteOptionsV4
+        let encodedData = NSMutableData()
+        let keyedArchiver = NSKeyedArchiver(forWritingWith: encodedData)
+        keyedArchiver.requiresSecureCoding = true
+        keyedArchiver.encode(options, forKey: "options")
+        keyedArchiver.finishEncoding()
+        
+        let keyedUnarchiver = NSKeyedUnarchiver(forReadingWith: encodedData as Data)
+        keyedUnarchiver.requiresSecureCoding = true
+        let unarchivedOptions = keyedUnarchiver.decodeObject(of: RouteOptionsV4.self, forKey: "options")!
+        keyedUnarchiver.finishDecoding()
+        
+        XCTAssertNotNil(unarchivedOptions)
+        
+        let coordinates = testCoordinates
+        let unarchivedWaypoints = unarchivedOptions.waypoints
+        XCTAssertEqual(unarchivedWaypoints.count, coordinates.count)
+        XCTAssertEqual(unarchivedWaypoints[0].coordinate.latitude, coordinates[0].latitude)
+        XCTAssertEqual(unarchivedWaypoints[0].coordinate.longitude, coordinates[0].longitude)
+        XCTAssertEqual(unarchivedWaypoints[1].coordinate.latitude, coordinates[1].latitude)
+        XCTAssertEqual(unarchivedWaypoints[1].coordinate.longitude, coordinates[1].longitude)
+        XCTAssertEqual(unarchivedWaypoints[2].coordinate.latitude, coordinates[2].latitude)
+        XCTAssertEqual(unarchivedWaypoints[2].coordinate.longitude, coordinates[2].longitude)
+        
+        XCTAssertEqual(unarchivedOptions.profileIdentifier, options.profileIdentifier)
+        XCTAssertEqual(unarchivedOptions.instructionFormat, options.instructionFormat)
+        XCTAssertEqual(unarchivedOptions.includesShapes, options.includesShapes)
+        XCTAssertEqual(unarchivedOptions.includesSteps, options.includesSteps)
+    }
+    
     func testCopying() {
-        let testInstance = RouteOptions.testInstance
+        let testInstance = testRouteOptions
         guard let copy = testInstance.copy() as? RouteOptions else { return XCTFail("RouteOptions copy method should an object of same type") }
         XCTAssertNotNil(copy, "Copy should not be nil.")
         XCTAssertTrue(testInstance == copy, "Test Instance and copy should be semantically equivalent.")
@@ -139,27 +170,32 @@ class RouteOptionsTests: XCTestCase {
     }
 }
 
-private extension RouteOptions {
-    static var testCoordinates: [CLLocationCoordinate2D] {
-        return [
-            CLLocationCoordinate2D(latitude: 52.5109, longitude: 13.4301),
-            CLLocationCoordinate2D(latitude: 52.5080, longitude: 13.4265),
-            CLLocationCoordinate2D(latitude: 52.5021, longitude: 13.4316),
-        ]
-    }
-    static var testInstance: RouteOptions {
-        let opts = RouteOptions(coordinates: self.testCoordinates, profileIdentifier: .automobileAvoidingTraffic)
-        opts.locale = Locale(identifier: "en_US")
-        opts.allowsUTurnAtWaypoint = true
-        opts.shapeFormat = .polyline
-        opts.routeShapeResolution = .full
-        opts.attributeOptions = [.congestionLevel]
-        opts.includesExitRoundaboutManeuver = true
-        opts.includesSpokenInstructions = true
-        opts.distanceMeasurementSystem = .metric
-        opts.includesVisualInstructions = true
-        opts.roadClassesToAvoid = .toll
-        
-        return opts
-    }
+let testCoordinates = [
+    CLLocationCoordinate2D(latitude: 52.5109, longitude: 13.4301),
+    CLLocationCoordinate2D(latitude: 52.5080, longitude: 13.4265),
+    CLLocationCoordinate2D(latitude: 52.5021, longitude: 13.4316),
+]
+
+var testRouteOptions: RouteOptions {
+    let opts = RouteOptions(coordinates: testCoordinates, profileIdentifier: .automobileAvoidingTraffic)
+    opts.locale = Locale(identifier: "en_US")
+    opts.allowsUTurnAtWaypoint = true
+    opts.shapeFormat = .polyline
+    opts.routeShapeResolution = .full
+    opts.attributeOptions = [.congestionLevel]
+    opts.includesExitRoundaboutManeuver = true
+    opts.includesSpokenInstructions = true
+    opts.distanceMeasurementSystem = .metric
+    opts.includesVisualInstructions = true
+    opts.roadClassesToAvoid = .toll
+    
+    return opts
+}
+
+var testRouteOptionsV4: RouteOptionsV4 {
+    let options = RouteOptionsV4(coordinates: testCoordinates, profileIdentifier: .automobileAvoidingTraffic)
+    options.instructionFormat = .html
+    options.includesShapes = true
+    options.includesSteps = true
+    return options
 }